chore: Renamed core/execution-context/execution-context.service.mts to core/execution-context/execution.mts.
