+++ packages/architectura/src/core/execution-context/_index.mts
@@ -1,2 +1,2 @@
+export * from './execution-context.mjs';
 export * from './execution-context.registry.mjs';
-export * from './execution-context.service.mjs';
